Fix EnrollmentEntity for database
EnrollmentEntity did not store the full uid string of the related class and person, it was relying on a join to the class and person entity. This fails when attempting to create a pending enrollment for a student who profile which is being created as part of a parental registration (on the client).